+++ src/Raven.Studio/jest.config.js
@@ -35,7 +35,9 @@ module.exports = {
         "^durandal/(.*)$": "<rootDir>/node_modules/durandal/js/$1",
         "^endpoints$":  "<rootDir>/typings/server/endpoints",
         "^d3$": "<rootDir>/wwwroot/Content/custom_d3",
+        "^hooks/(.*)$": "<rootDir>/typescript/components/hooks/$1",
         "\\.(css|less|scss)$": "<rootDir>/typescript/test/__mocks__/styleMock.js",
+        "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga|docx|pdf)$": "<rootDir>/typescript/test/__mocks__/fileMock.js",
     },
     "reporters": [
         "default",
