+++ src/Raven.Studio/typescript/components/pages/database/indexes/list/IndexGlobalIndexing.tsx
