+++ src/Raven.Studio/typescript/components/pages/database/indexes/list/IndexProgress.tsx
@@ -1,5 +1,5 @@
 ﻿import React from "react";
-import { IndexCollectionProgress, IndexNodeInfoDetails, IndexProgressInfo, Progress } from "../../../models/indexes";
+import { IndexCollectionProgress, IndexNodeInfoDetails, IndexProgressInfo, Progress } from "../../../../models/indexes";
 import classNames from "classnames";
 import IndexRunningStatus = Raven.Client.Documents.Indexes.IndexRunningStatus;
 import genUtils from "common/generalUtils";
