+++ src/Raven.Studio/typescript/viewmodels/database/indexes/indexes.ts
@@ -1,5 +1,5 @@
 import shardViewModelBase from "viewmodels/shardViewModelBase";
-import { IndexesPage } from "../../../components/pages/database/indexes/IndexesPage";
+import { IndexesPage } from "../../../components/pages/database/indexes/list/IndexesPage";
 
 export class indexes extends shardViewModelBase {
     view = { default: `<div class="indexes content-margin no-transition absolute-fill" data-bind="react: reactOptions"></div>` };
