+++ src/Raven.Studio/wwwroot/Content/scss/_colors.scss
