+++ src/Raven.Studio/wwwroot/Content/scss/_sizes.scss
