+++ src/Raven.Studio/wwwroot/Content/scss/_variables.scss
